+++ semcore/date-picker/__tests__/date-picker.browser-test.tsx
@@ -588,7 +588,6 @@ test.describe(`${TAG.FUNCTIONAL}`, () => {
         await locators.button(page, 'Previous month').waitFor({ state: 'visible' });
         await page.keyboard.press('ArrowDown');
         const newValue2 = await input.inputValue();
-
         await page.keyboard.press('Tab');
         await expect(locators.button(page, 'Today')).toBeFocused();